Get rid of kinto.core.logs module
