+++ kinto/core/utils.py
@@ -262,6 +262,16 @@ def reapply_cors(request, response):
     return response
 
 
+def log_context(request, **kwargs):
+    """Bind information to the current request summary log.
+    """
+    try:
+        request._log_context.update(**kwargs)
+    except AttributeError:
+        request._log_context = kwargs
+    return request._log_context
+
+
 def current_service(request):
     """Return the Cornice service matching the specified request.
